+++ bitcoin/src/blockdata/constants.rs
@@ -30,7 +30,7 @@ pub const DIFFCHANGE_INTERVAL: u32 = 2016;
 pub const DIFFCHANGE_TIMESPAN: u32 = 14 * 24 * 3600;
 
 /// The factor that non-witness serialization data is multiplied by during weight calculation.
-pub const WITNESS_SCALE_FACTOR: usize = 4;
+pub const WITNESS_SCALE_FACTOR: usize = units::weight::WITNESS_SCALE_FACTOR;
 /// The maximum allowed number of signature check operations in a block.
 pub const MAX_BLOCK_SIGOPS_COST: i64 = 80_000;
 /// Mainnet (bitcoin) pubkey address prefix.
