+++ bitcoin/src/lib.rs
@@ -89,7 +89,6 @@ extern crate actual_serde as serde;
 #[macro_use]
 mod test_macros;
 mod internal_macros;
-mod parse;
 #[cfg(feature = "serde")]
 mod serde_utils;
 
@@ -196,4 +195,10 @@ pub mod amount {
             self.to_sat().consensus_encode(w)
         }
     }
+}
+
+/// Unit parsing utilities.
+pub mod parse {
+    /// Re-export everything from the [`units::parse`] module.
+    pub use units::parse::ParseIntError;
 }
