+++ units/src/locktime/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: CC0-1.0
+
+//! Provides absolute and relative locktimes.
+
+pub mod absolute;
+pub mod relative;
